Defining Services type and setting services functions to call directly crossFetch lib for production mode
