+++ packages/@coorpacademy-app-review/src/services/index.ts
@@ -1,7 +1,8 @@
+import {Services} from '../types/common';
 import {fetchSkills} from './fetch-skills';
 import {postProgression} from './post-progression';
 
-export const getServices = () => ({
+export const getServices = (): Services => ({
   fetchSkills,
   postProgression
 });
